backend/functions/on_profile_create/requirements.txt
google-cloud-firestore==2.19.0
anthropic==0.42.0
functions-framework==3.8.1
